requirements.txt
numpy
scipy
scikit-learn
scikit-image
pandas
matplotlib
imageio
opencv-python
opencv-contrib-python 
pyzmq
torch
torchvision